+++ projects/angular4-paystack/src/lib/angular4-paystack-embed.component.spec.ts
@@ -1,16 +1,16 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { Angular4PaystackComponent } from './angular4-paystack.component';
+import { Angular4PaystackEmbedComponent } from './angular4-paystack-embed.component';
 import { Angular4PaystackService } from './angular4-paystack.service';
 import { PUBLIC_KEY_TOKEN } from './paystack-token';
 
-describe('Angular4PaystackComponent', () => {
-  let component: Angular4PaystackComponent;
-  let fixture: ComponentFixture<Angular4PaystackComponent>;
+describe('Angular4PaystackEmbedComponent', () => {
+  let component: Angular4PaystackEmbedComponent;
+  let fixture: ComponentFixture<Angular4PaystackEmbedComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ Angular4PaystackComponent ],
+      declarations: [ Angular4PaystackEmbedComponent ],
       providers: [
         Angular4PaystackService,
         { provide: PUBLIC_KEY_TOKEN, useValue: 'public-key' }
@@ -20,11 +20,14 @@ describe('Angular4PaystackComponent', () => {
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(Angular4PaystackComponent);
+    fixture = TestBed.createComponent(Angular4PaystackEmbedComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
 
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
   it('should create', () => {
     expect(component).toBeTruthy();
   });