+++ projects/angular4-paystack/src/lib/angular4-paystack.directive.ts
@@ -1,5 +1,6 @@
 import { Directive, Input, Output, EventEmitter, HostListener } from '@angular/core';
-import { PaystackOptions } from './paystack-options';
+import { PaystackOptions, PrivatePaystackOptions } from './paystack-options';
+import { Angular4PaystackService } from './angular4-paystack.service';
 
 interface MyWindow extends Window {
   PaystackPop: {
@@ -9,10 +10,9 @@ interface MyWindow extends Window {
 declare var window: MyWindow;
 
 @Directive({
-  selector: '[angular4-paystack]',
+  selector: '[angular4-paystack]', // tslint:disable-line
 })
 export class Angular4PaystackDirective {
-  @Input() text: string;
   @Input() key: string;
   @Input() email: string;
   @Input() amount: number;
@@ -23,84 +23,63 @@ export class Angular4PaystackDirective {
   @Input() quantity: string;
   @Input() subaccount: string;
   @Input() channels: string[];
-  @Input() transaction_charge: number;
+  @Input() transaction_charge: number; // tslint:disable-line
   @Input() bearer: string;
   @Input() class: string;
   @Input() style: object;
-  @Output() close: EventEmitter<string> = new EventEmitter<string>();
-  @Output() callback: EventEmitter<string> = new EventEmitter<string>();
-  private paystackOptions: PaystackOptions;
+  @Input() paystackOptions: PaystackOptions;
+  @Output() paymentInit: EventEmitter<any> = new EventEmitter<any>(); // tslint:disable-line
+  @Output() onClose: EventEmitter<any> = new EventEmitter<any>(); // tslint:disable-line
+  @Output() callback: EventEmitter<any> = new EventEmitter<any>();
+  private _paystackOptions: Partial<PrivatePaystackOptions>; // tslint:disable-line
   private isPaying = false;
-  constructor() {
-    this.setUp();
-  }
 
-  pay() {
-    this.setUp();
-    if (!this.checkInput()) { return; }
-    const payment = window.PaystackPop.setup(this.paystackOptions);
+  constructor(private paystackService: Angular4PaystackService) {}
+
+  async pay() {
+    let errorText = '';
+    if (this.paystackOptions && Object.keys(this.paystackOptions).length >= 2) {
+      errorText = this.valdateInput(this.paystackOptions);
+      this.generateOptions(this.paystackOptions);
+    } else {
+      errorText = this.valdateInput(this);
+      this.generateOptions(this);
+    }
+    if (errorText) {
+      console.error(errorText);
+      return errorText;
+    }
+    await this.paystackService.loadScript();
+    if (this.isPaying) { return; }
+    if (this.paymentInit.observers.length) {
+      this.paymentInit.emit();
+    }
+    const payment = window.PaystackPop.setup(this._paystackOptions);
     payment.openIframe();
     this.isPaying = true;
   }
-  checkInput() {
-    if (!this.key) { return console.error('ANGULAR-PAYSTACK: Paystack key cannot be empty'); }
-    if (!this.email) { return console.error('ANGULAR-PAYSTACK: Paystack email cannot be empty'); }
-    if (!this.amount) { return console.error('ANGULAR-PAYSTACK: Paystack amount cannot be empty'); }
-    if (!this.ref) { return console.error('ANGULAR-PAYSTACK: Paystack ref cannot be empty'); }
+
+  valdateInput(obj: PaystackOptions) {
     if (!this.callback.observers.length) {
-      return console.error(
-        `ANGULAR-PAYSTACK: Insert a callback output like so (callback)='PaymentComplete($event)' to check payment status`
-      );
+      return 'ANGULAR-PAYSTACK: Insert a callback output like so (callback)=\'PaymentComplete($event)\' to check payment status';
     }
-    return true;
+    return this.paystackService.checkInput(obj);
   }
 
-  setUp() {
-    this.paystackOptions = {
-      key: this.key ,
-      email: this.email ,
-      amount: this.amount ,
-      ref: this.ref ,
-      metadata: this.metadata || {},
-      currency: this.currency || 'NGN' ,
-      plan: this.plan || '' ,
-      quantity: this.quantity || '' ,
-      subaccount: this.subaccount || '' ,
-      channels: this.channels,
-      transaction_charge: this.transaction_charge || 0 ,
-      bearer: this.bearer || '' ,
-      callback: (res) => {
-        this.isPaying = false;
-        this.callback.emit(res);
-      },
-      onClose: () => {
-        this.isPaying = false;
-        this.close.emit();
+  generateOptions(obj: PaystackOptions) {
+    this._paystackOptions = this.paystackService.getPaystackOptions(obj);
+    this._paystackOptions.onClose = () => {
+      if (this.onClose.observers.length) {
+        this.onClose.emit();
       }
     };
-  }
-
-  loadScript(): Promise<void> {
-    return new Promise(resolve => {
-      if (window.PaystackPop && typeof window.PaystackPop.setup === 'function') {
-        resolve();
-        return;
-      }
-      const script = window.document.createElement('script');
-      window.document.head.appendChild(script);
-      const onLoadFunc = () => {
-        script.removeEventListener('load', onLoadFunc);
-        resolve();
-      };
-      script.addEventListener('load', onLoadFunc);
-      script.setAttribute('src', 'https://js.paystack.co/v1/inline.js');
-    });
+    this._paystackOptions.callback = (...response) => {
+      this.callback.emit(...response);
+    };
   }
 
   @HostListener('click')
   async buttonClick() {
-    if (this.isPaying) { return; }
-    await this.loadScript();
     this.pay();
   }
 }