+++ projects/angular4-paystack/src/lib/paystack-token.ts
@@ -1,0 +1,3 @@
+import { InjectionToken } from '@angular/core';
+
+export const PUBLIC_KEY_TOKEN = new InjectionToken<string>('paystack.publickey');